+++ src/Exceptions/InvalidComponentDataException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Storyblok\ManagementApi\Exceptions;
+
+/**
+ * Exception thrown when invalid story data is provided
+ */
+class InvalidComponentDataException extends \InvalidArgumentException {}