+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
+## 1.0.2 - WIP
+- Implementing ComponentApi class for handling components
+
 ## 1.0.1 - 2025-05-09
 - Fixing getting list of Stories. Issue #10
 - Adding new plans in `StoryblokUtils::getPlanDescription()` method
 - Reviewing README
-- Adding dependecies license checker
+- Adding dependencies license checker
 
 ## 1.0.0 - 2025-03-14
 - Releasing v 1.0.0